Split reader into modules
